+++ src/main/kotlin/com/github/kangwooc/dbmlplugin/lang/DBMLLanguage.kt
@@ -1,0 +1,7 @@
+package com.github.kangwooc.dbmlplugin.lang
+
+import com.intellij.lang.Language
+
+object DBMLLanguage : Language("DBML") {
+    private fun readResolve(): Any = DBMLLanguage
+}